+++ migration-library/src/test/kotlin/com/quadient/migration/service/inspirebuilder/InspireDocumentObjectBuilderTest.kt
@@ -570,7 +570,7 @@ class InspireDocumentObjectBuilderTest {
         val cell = result["Cell"].last { it["Id"].textValue() == tableRowSet["SubRowId"].textValue() }
         val flow = result["Flow"].last { it["Id"].textValue() == cell["FlowId"].textValue() }
         // first paragraph = warning, second paragraph = original "Name"
-        flow["FlowContent"]["P"][0]["T"][""].textValue().shouldBeEqualTo($$"<repeated by unmapped $Clients$>")
+        flow["FlowContent"]["P"][0]["T"][""].textValue().shouldBeEqualTo($$"<repeated row by unmapped $Clients$>")
         flow["FlowContent"]["P"][1]["T"][""].textValue().shouldBeEqualTo("Name")
     }
 
@@ -601,7 +601,7 @@ class InspireDocumentObjectBuilderTest {
         val firstCell = result["Cell"].last { it["Id"].textValue() == firstSingleRow["SubRowId"].textValue() }
         val firstFlow = result["Flow"].last { it["Id"].textValue() == firstCell["FlowId"].textValue() }
         firstFlow["FlowContent"]["P"][0]["T"][""].textValue()
-            .shouldBeEqualTo($$"<repeated by unmapped $Data.Clients.Value$>")
+            .shouldBeEqualTo($$"<repeated row by unmapped $Data.Clients.Value$>")
         firstFlow["FlowContent"]["P"][1]["T"][""].textValue().shouldBeEqualTo($$"$surname$")
 
         val secondSingleRow =
@@ -696,7 +696,7 @@ class InspireDocumentObjectBuilderTest {
         repeatedFallbackFlow["Type"].textValue().shouldBeEqualTo("Simple")
         repeatedFallbackFlow["SectionFlow"].textValue().shouldBeEqualTo("False")
         repeatedFallbackFlow["FlowContent"]["P"][0]["T"][""].textValue()
-            .shouldBeEqualTo($$"<repeated by unmapped $Clients$>")
+            .shouldBeEqualTo($$"<repeated content by unmapped $Clients$>")
         repeatedFallbackFlow["FlowContent"]["P"][1]["T"][""].textValue().shouldBeEqualTo("Some content")
     }